+++ src/constants/message.ts
@@ -175,57 +175,3 @@ export const INFO_MESSAGES = {
   REVIEW_INFO: '서비스 완료 후 리뷰를 작성할 수 있습니다.',
   MATCHING_INFO: '선호하는 매니저를 선택하거나 자동 매칭을 이용할 수 있습니다.',
 } as const;
-
-// ===== 버튼 텍스트 =====
-export const BUTTON_TEXTS = {
-  // 기본 액션
-  SAVE: '저장',
-  CANCEL: '취소',
-  DELETE: '삭제',
-  EDIT: '수정',
-  CREATE: '생성',
-  UPDATE: '업데이트',
-  SUBMIT: '제출',
-  RESET: '초기화',
-
-  // 인증
-  LOGIN: '로그인',
-  LOGOUT: '로그아웃',
-  SIGNUP: '회원가입',
-
-  // 예약
-  RESERVE: '예약하기',
-  CANCEL_RESERVATION: '예약 취소',
-  APPROVE: '승인',
-  REJECT: '거절',
-
-  // 찜하기/블랙리스트
-  LIKE: '찜하기',
-  UNLIKE: '찜 해제',
-  BLACKLIST: '블랙리스트',
-
-  // 리뷰
-  WRITE_REVIEW: '리뷰 작성',
-  SUBMIT_REVIEW: '리뷰 제출',
-
-  // 네비게이션
-  BACK: '뒤로',
-  NEXT: '다음',
-  PREVIOUS: '이전',
-  HOME: '홈',
-
-  // 검색/필터
-  SEARCH: '검색',
-  FILTER: '필터',
-  RESET_FILTER: '필터 초기화',
-
-  // 파일
-  UPLOAD: '업로드',
-  DOWNLOAD: '다운로드',
-
-  // 기타
-  CONFIRM: '확인',
-  CLOSE: '닫기',
-  VIEW_MORE: '더보기',
-  REFRESH: '새로고침',
-} as const;
